refactor(data-loader): refactor usecase layer
